compmbox: fix long-to-int truncation in comp_mbox_check()
mutt_file_get_size() returns long and CompressInfo::size is long, but
the local variable storing the result was declared as int. For
compressed files larger than 2GB, this truncation would cause incorrect
size comparisons — potentially missing new mail or triggering
unnecessary decompression.

Change the local variable type from int to long.

+++ compmbox/compress.c
@@ -562,7 +562,7 @@ static enum MxStatus comp_mbox_check(struct Mailbox *m)
   if (!ops)
     return MX_STATUS_ERROR;
 
-  int size = mutt_file_get_size(m->realpath);
+  long size = mutt_file_get_size(m->realpath);
   if (size == ci->size)
     return MX_STATUS_OK;
